change the type of leverage

+++ programs/ostium/src/account.rs
@@ -17,7 +17,7 @@ pub struct Position {
     pub exit_price: u64,            // the asset price when this position was closed
     pub exit_timestamp: u64,        // the timestamp when this position was closed
     pub quantity: u64, // the asset quantity, with no leverage. The smaller unit is 0.0001. (ex: 1 ounce of gold will be 10000, 0.5 ounce of gold will be 5000)
-    pub leverage: u16, // if the leverage is 10 and the quantity is 0.5, it means that we are actually trading with 5 ounces of gold (0.5 from the user, 4.5 borrowed).
+    pub leverage: u64, // if the leverage is 10 and the quantity is 0.5, it means that we are actually trading with 5 ounces of gold (0.5 from the user, 4.5 borrowed).
     pub pos_type: PositionType, // type of this position
     pub pos_status: PositionStatus, // status of this position
 }
